state test splitted into different files
