+++ elements/itemfilter/test/cases/state/index.js
@@ -1,0 +1,6 @@
+// Exported test methods
+
+export { default as renderCheckboxTest } from "./render-checkbox";
+export { default as resultsWithPreSetFiltersTest } from "./results-with-pre-set-filters";
+export { default as resultPreCheckedTest } from "./result-pre-checked";
+export { default as customOrderTest } from "./custom-order";
