+++ elements/itemfilter/test/cases/state/render-checkbox.js
@@ -1,0 +1,13 @@
+const renderCheckboxTest = () => {
+  cy.get("eox-itemfilter")
+    .shadow()
+    .within(() => {
+      cy.get("eox-itemfilter-select")
+        .shadow()
+        .within(() => {
+          cy.get('[type="checkbox"]:checked').should("have.length", 2);
+        });
+    });
+};
+
+export default renderCheckboxTest;
